refactor: 상품 등록 화면 titleLabel 강제 언래핑 개선 #4
- Double+Extension 파일 생성

+++ OpenMarket/OpenMarket/Extension/Double+Extension.swift
@@ -1,6 +1,6 @@
 import Foundation
 
-extension Int {
+extension Double {
     func formattedWithComma() -> String {
         let numberFormatter = NumberFormatter()
         numberFormatter.numberStyle = .decimal
